some compute project cleanup / file breakup
