+++ compute/NodePen.Compute.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2010
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30907.101
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "resthopper", "resthopper.csproj", "{01757ED0-58C4-47DA-926A-70ACD31B93E1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NodePen.Compute", "NodePen.Compute.csproj", "{01757ED0-58C4-47DA-926A-70ACD31B93E1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,10 +15,6 @@ Global
 		{01757ED0-58C4-47DA-926A-70ACD31B93E1}.Debug|x64.Build.0 = Debug|x64
 		{01757ED0-58C4-47DA-926A-70ACD31B93E1}.Release|x64.ActiveCfg = Release|x64
 		{01757ED0-58C4-47DA-926A-70ACD31B93E1}.Release|x64.Build.0 = Release|x64
-		{5AE28138-B0A2-40C1-B62C-67CC7371E992}.Debug|x64.ActiveCfg = Debug|x64
-		{5AE28138-B0A2-40C1-B62C-67CC7371E992}.Debug|x64.Build.0 = Debug|x64
-		{5AE28138-B0A2-40C1-B62C-67CC7371E992}.Release|x64.ActiveCfg = Release|x64
-		{5AE28138-B0A2-40C1-B62C-67CC7371E992}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
